# lab2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Приют и передача животных в семью",
    "description": "REST API приюта: животные, вольеры, заявки на усыновление, акты передачи, пользователи. Даты и время проставляет сервер, в запросах не передаются.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/animal-handover-records": {
      "get": {
        "tags": [
          "Акты передачи"
        ],
        "summary": "Список актов передачи",
        "operationId": "listHandoverRecords",
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHandoverRecordResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Акты передачи"
        ],
        "summary": "Подготовить акт передачи",
        "operationId": "createHandoverRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalHandoverRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Ресурс создан",
            "headers": {
              "Location": {
                "description": "Адрес созданного ресурса",
                "style": "simple"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalHandoverRecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/animals/{id}": {
      "get": {
        "tags": [
          "Животные"
        ],
        "summary": "Животное по идентификатору",
        "operationId": "getAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Животные"
        ],
        "summary": "Удалить животное",
        "operationId": "deleteAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Животные"
        ],
        "summary": "Частично обновить животное",
        "operationId": "updateAnimal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/animal-handover-records/{id}": {
      "get": {
        "tags": [
          "Акты передачи"
        ],
        "summary": "Акт передачи по идентификатору",
        "operationId": "getHandoverRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalHandoverRecordResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Акты передачи"
        ],
        "summary": "Удалить акт передачи",
        "operationId": "deleteHandoverRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Акты передачи"
        ],
        "summary": "Изменить статус акта передачи",
        "operationId": "updateHandoverRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalHandoverRecordPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalHandoverRecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Тело ответа с ошибкой",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Момент формирования ответа",
            "example": "2026-09-26T07:07:27.667885Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Код ответа",
            "example": 409
          },
          "error": {
            "type": "string",
            "description": "Причина, соответствующая коду",
            "example": "Conflict"
          },
          "message": {
            "type": "string",
            "description": "Описание причины ошибки"
          },
          "path": {
            "type": "string",
            "description": "Путь запроса",
            "example": "/api/animals/1"
          }
        }
      },
...
      "AnimalResponse": {
        "type": "object",
        "description": "Животное приюта",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Кличка",
            "example": "Бобик"
          },
          "species": {
            "type": "string",
            "description": "Вид животного",
            "example": "Собака"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Возраст в полных годах",
            "example": 3
          },
          "status": {
            "type": "string",
            "description": "Статус размещения",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "HANDED_OVER"
            ],
            "example": "AVAILABLE"
          }
        }
      },
      "AnimalHandoverRecordRequest": {
        "type": "object",
        "description": "Заявка, по которой готовится акт передачи",
        "properties": {
          "applicationId": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор одобренной заявки",
            "example": 1
          }
        },
        "required": [
          "applicationId"
        ]
      },
      "AnimalHandoverRecordResponse": {
        "type": "object",
        "description": "Акт передачи животного семье",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "applicationId": {
            "type": "integer",
            "format": "int64",
            "description": "Заявка, по которой оформлена передача",
            "example": 1
          },
          "preparationDate": {
            "type": "string",
            "format": "date",
            "description": "Дата подготовки акта, проставляется сервером",
            "example": "2026-09-26"
          },
          "confirmationDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Дата подтверждения акта, проставляется сервером при подтверждении",
            "example": "2026-09-26"
          },
          "status": {
            "type": "string",
            "description": "Статус акта передачи",
            "enum": [
              "PREPARED",
              "CONFIRMED",
              "CANCELLED"
            ],
            "example": "PREPARED"
          }
        }
      },
...
      "AnimalPatch": {
        "type": "object",
        "description": "Поля животного для изменения, отсутствующие поля не меняются",
        "properties": {
          "name": {
            "type": "string",
            "description": "Кличка",
            "example": "Бобик",
            "pattern": "\\S+"
          },
          "species": {
            "type": "string",
            "description": "Вид животного",
            "example": "Собака",
            "pattern": "\\S+"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Возраст в полных годах",
            "example": 4
          },
          "status": {
            "type": "string",
            "description": "Статус размещения",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "HANDED_OVER"
            ],
            "example": "AVAILABLE"
          }
        }
      },
      "AnimalHandoverRecordPatch": {
        "type": "object",
        "description": "Новый статус акта передачи",
        "properties": {
          "status": {
            "type": "string",
            "description": "Статус акта передачи",
            "enum": [
              "PREPARED",
              "CONFIRMED",
              "CANCELLED"
            ],
            "example": "CONFIRMED"
          }
        }
      },
...
    }
  }
}